FED related code

+++ web/portal/src/components/EditPanel/EditQuestion/index.js
@@ -64,7 +64,8 @@ class EditQuestion extends PureComponent {
                             {this._renderTitle()}
                             {this._renderRequired()}
                             {this._renderOptions()}
-                            {editQuestion.type === 'rating' ? this._renderAdvance() : ''}
+                            {editQuestion.type === 'rating' || editQuestion.type === 'textarea' ?
+                                this._renderAdvance() : ''}
                         </div>
                     </div>
                     <div className="bottom">
@@ -98,7 +99,7 @@ class EditQuestion extends PureComponent {
                 <div>
                     <textarea
                         id="editQuestion"
-                        className="ut-editQuestion teaxarea"
+                        className="ut-editQuestion textarea"
                         type="text"
                         value={editQuestion.label}
                         onChange={this._onTitleChange}
@@ -133,10 +134,13 @@ class EditQuestion extends PureComponent {
 
     _renderType() {
         const { editQuestion } = this.props;
+        // Once add new question, it should add a select option
         const item = [
             { value: 'radio', label: 'Radio Button' },
             { value: 'checkbox', label: 'Checkbox' },
-            { value: 'rating', label: 'Rating (Likert Scale)' }];
+            { value: 'rating', label: 'Rating (Likert Scale)' },
+            { value: 'text', label: 'Single line text' },
+            { value: 'textarea', label: 'Multiple line text' }];
         return (
             <div className={styles.editSection}>
                 <div className={styles.title}>Question Type</div>
@@ -158,10 +162,12 @@ class EditQuestion extends PureComponent {
             editQuestion,
             handleChangeEvent: this._handleChangeEvent
         };
+        const sectionTitle = editQuestion.type === 'text' || editQuestion.type === 'textarea' ?
+            'Placeholder Settings' : 'Multiple Choice Options';
 
         return (
             <div className={styles.editSection}>
-                <div className={styles.title}>Multiple Choice Options</div>
+                <div className={styles.title}>{sectionTitle}</div>
                 {<EditMultiOptions {...props} />}
             </div>
         );
@@ -169,43 +175,70 @@ class EditQuestion extends PureComponent {
 
     _renderAdvance() {
         const { editQuestion } = this.props;
-        const flag = editQuestion.hasOwnProperty('input');
-        const input = flag ? editQuestion.input : values.PLACEHOLDER_TITLE;
 
-        return (
-            <div className={`${styles.editSection} ut-advance`}>
-                <div className={styles.title}>Advanced Option Settings</div>
+        if (editQuestion.type === 'rating') {
+            const flag = editQuestion.hasOwnProperty('input');
+            const input = flag ? editQuestion.input : values.PLACEHOLDER_TITLE;
 
-                <div className={`${styles.checkboxItem} checkboxItem`}>
-                    <input
-                        id="chk"
-                        type="checkbox"
-                        className={`${styles.input} ut-chk input`}
-                        checked={flag}
-                        onChange={this._onAdvanceChangeHandle}
-                    />
-                    <label>
-                        Show "Tell Me Why"<span>&nbsp;-&nbsp;</span>
-                    </label>
-                    <input
-                        id="why"
-                        type="text"
-                        value={input}
-                        className={`${styles.input__why} input input--medium`}
-                        onChange={this._onAdvanceChangeHandle}
-                        onFocus={this._handleFocusEvent}
-                        readOnly={!flag}
-                    />
-                    <div className={`${styles.whyInput} input__msg js-whyInput-msg`}></div>
+            return (
+                <div className={`${styles.editSection} ut-advance`}>
+                    <div className={styles.title}>Advanced Option Settings</div>
+
+                    <div className={`${styles.checkboxItem} checkboxItem`}>
+                        <input
+                            id="chk"
+                            type="checkbox"
+                            className={`${styles.input} ut-chk input`}
+                            checked={flag}
+                            onChange={this._onAdvanceChangeHandle}
+                        />
+                        <label>
+                            Show "Tell Me Why"<span>&nbsp;-&nbsp;</span>
+                        </label>
+                        <input
+                            id="why"
+                            type="text"
+                            value={input}
+                            className={`${styles.input__why} input input--medium`}
+                            onChange={this._onAdvanceChangeHandle}
+                            onFocus={this._handleFocusEvent}
+                            readOnly={!flag}
+                        />
+                        <div className={`${styles.whyInput} input__msg js-whyInput-msg`}></div>
+                    </div>
                 </div>
-            </div>
-        );
+            );
+        } else if (editQuestion.type === 'textarea') {
+            return (
+                <div className={`${styles.editSection} ut-advance`}>
+                    <div className={styles.title}>Advanced Settings</div>
+                    <div>
+                        <div className={styles.rows}>Show rows:</div>
+                        <button
+                            data-type="decrease"
+                            onClick={this._onAdvanceChangeHandle}
+                        >-</button>
+                        <input
+                            id="number"
+                            type="text"
+                            className={`${styles.input__num} input input--medium`}
+                            value={editQuestion.rows}
+                            readOnly
+                        />
+                        <button
+                            data-type="increase"
+                            onClick={this._onAdvanceChangeHandle}
+                        >+</button>
+                    </div>
+                </div>
+            );
+        }
     }
 
     _btnClickEvent(e) {
         const { editQuestion, questionsActions, editQuestionActions } = this.props;
         // Clear all error msg
-        $('.js-title-msg, .js-opt-msg, .js-optInput-msg, .js-whyInput-msg').html('');
+        $('.js-title-msg, .js-opt-msg, .js-optInput-msg, .js-whyInput-msg, js-text-msg').html('');
 
         if (e.currentTarget.getAttribute('data-type') === 'cancel') {
             editQuestionActions.stopEditQuestion();
@@ -243,6 +276,14 @@ class EditQuestion extends PureComponent {
                 $('.js-whyInput-msg').html('Please fill "Tell Me Why" input\'s placeholder');
                 flag = true;
             }
+            // Check single or multi line text
+            opt = $('.js-text');
+            if (opt && opt.length > 0) {
+                if (opt[0].value === '') {
+                    $('.js-text-msg').html('Please fill placehoder');
+                    flag = true;
+                }
+            }
 
             if (!flag) {
                 // save editQuestion to Question
@@ -267,21 +308,34 @@ class EditQuestion extends PureComponent {
             type,
             label: editQuestion.label,
             order: editQuestion.order,
-            required: editQuestion.required,
-            data: []
+            required: editQuestion.required
         });
+        // Once add new question, it should add case content
         // keep advanced question content by question type
         switch (type) {
         case 'radio':
         case 'checkbox':
-            newQuestion.data = [...editQuestion.data];
+            newQuestion.data = !editQuestion.hasOwnProperty('data') ?
+                [{ value: Mixins.generateQuestionID(), label: values.OPTION_TITLE }] :
+                [...editQuestion.data];
             break;
         case 'rating':
-            newQuestion.data = [...editQuestion.data];
-            // rating's options should not have input
-            editQuestion.data.forEach((opt, idx) => {
-                if (opt.hasOwnProperty('input')) delete newQuestion.data[idx].input;
-            });
+            newQuestion.data = !editQuestion.hasOwnProperty('data') ?
+                [{ value: Mixins.generateQuestionID(), label: values.OPTION_TITLE }] :
+                [...editQuestion.data];
+            if (editQuestion.hasOwnProperty('data')) {
+                // rating's options should not have input
+                editQuestion.data.forEach((opt, idx) => {
+                    if (opt.hasOwnProperty('input')) delete newQuestion.data[idx].input;
+                });
+            }
+            break;
+        case 'text':
+            newQuestion.input = values.PLACEHOLDER_TITLE;
+            break;
+        case 'textarea':
+            newQuestion.input = values.PLACEHOLDER_TITLE;
+            newQuestion.rows = 3;
             break;
         default:
         }
@@ -293,16 +347,30 @@ class EditQuestion extends PureComponent {
         editQuestionActions.setEditQuestion(data);
     }
 
-    _onAdvanceChangeHandle() {
+    _onAdvanceChangeHandle(e) {
         const { editQuestion, editQuestionActions } = this.props;
-        const flag = $('#chk').is(':checked');
-        const newData = Object.assign({}, editQuestion);
-        if (flag) {
-            newData.input = $('#why').val();
+        const target = e.target;
+        if (editQuestion.type === 'rating') {
+            const flag = $('#chk').is(':checked');
+            const newData = Object.assign({}, editQuestion);
+            if (flag) {
+                newData.input = $('#why').val();
+                this._handleChangeEvent(newData);
+            } else {
+                delete newData.input;
+                editQuestionActions.deleteRatingInput();
+            }
+        } else if (editQuestion.type === 'textarea') {
+            let number = parseInt($('#number').val(), 10);
+            if (target.getAttribute('data-type') === 'increase') {
+                number++;
+                number = number > 10 ? 10 : number;
+            } else {
+                number--;
+                number = number < 2 ? 2 : number;
+            }
+            const newData = Object.assign({}, editQuestion, { rows: number });
             this._handleChangeEvent(newData);
-        } else {
-            delete newData.input;
-            editQuestionActions.deleteRatingInput();
         }
     }
